BasicTests

+++ kodilla-collections/src/main/java/com/kodilla/collections/interfaces/homework/Bmw.java
@@ -2,22 +2,22 @@ package com.kodilla.collections.interfaces.homework;
 
 public class Bmw implements Car{
 
-    int speed;
-    private int acceleration;
-    private int brakes;
+    double speed;
+    private double acceleration;
+    private double brakes;
 
-    public Bmw(double speed, double acceleration, double brakes) {
+    public Bmw(int speed, int acceleration, int brakes) {
         this.speed = speed;
         this.acceleration = acceleration;
         this.brakes = brakes;
     }
 
-    public int getSpeed() {
+    public double getSpeed() {
 
         return speed;
     }
 
-    public int getAcceleration() {
+    public double getAcceleration() {
         return acceleration;
     }
