+++ kodilla-collections/src/main/java/com/kodilla/collections/interfaces/homework/Mazda.java
@@ -2,21 +2,21 @@ package com.kodilla.collections.interfaces.homework;
 
 public class Mazda implements Car{
 
-    int speed;
-    private int acceleration;
-    private int brakes;
+    double speed;
+    private double acceleration;
+    private double brakes;
 
-    public Mazda(double speed, double acceleration, double brakes) {
+    public Mazda(int speed, int acceleration, int brakes) {
         this.speed = speed;
         this.acceleration = acceleration;
         this.brakes = brakes;
     }
-    public int getSpeed() {
+    public double getSpeed() {
 
         return speed;
     }
 
-    public int getAcceleration() {
+    public double getAcceleration() {
         return acceleration;
     }
